\name{BicatYeast}
\alias{BicatYeast}
\alias{yeast microarray reduced data matrix}
\alias{Saccharomices Cerevisiae reduced data matrix}

\docType{data}

\title{BicAT Yeast}
\description{Microarray data matrix for 80 experiments with Saccharomyces Cerevisiae organism
  extracted from BicAT example data set.}
\usage{
  data(BicatYeast)
  }

\format{
  Data structure with information about the expression levels of 419 probesets 
  over 70 conditions
  Row names follow Affymetrix probeset notation
  }
%\details{}
%\value{}
%\references{Si acaso Eisen1998}


\source{BicAT datasets at http://www.tik.ee.ethz.ch/sop/bicat/
}
%\seealso{
  %parallelCoordinates for alternative representation of biclusters.
%  \item{\code{\link{bubbleplot}}}{ for simultaneous representation of biclusters.}
%  }
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{datasets}
