\name{biclust-package}
\alias{biclust-package}
\alias{biclust}
\docType{package}
\title{
Performing state of the art bicluster algorithms
}
\description{}
\details{
\tabular{ll}{
Package: \tab biclust\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2007-05-03\cr
License: \tab GPL 2 or higher\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{Sebastian Kaiser

Maintainer:Sebastian Kaiser <sebastian.kaiser@stat.uni-muenchen.de>
}
\references{}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{bicluster}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
