% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anRpackage-internal.R
\name{diagnosticPlot2}
\alias{diagnosticPlot2}
\title{Diagnostics F Statistiics Visualization}
\usage{
diagnosticPlot2(diagnosticTest, number = 1, StatVal = TRUE,
  binwidth = NULL)
}
\arguments{
\item{diagnosticTest}{output of \code{\link{diagnosticTest}} with \code{save_F=TRUE} which contains the F-statistics and sampling replicates.}

\item{number}{Number of which BC to plot. This needs to be one of the Biclusters requested in in \code{\link{diagnosticTest}}.}

\item{StatVal}{Boolean value to draw the observed statistic on the distribution plots.}

\item{binwidth}{The width of the bins.}
}
\value{
Returns a \code{ggplot} object.
}
\description{
Plots distributions of bootstrap replicates of F-statistics for row, column and multiplicative effects obtained from \code{\link{diagnosticTest}} (when \code{save_F=TRUE}). 
Contains an option to highlight the observed statistics.
}
\examples{

#Random matrix with embedded bicluster
test <- matrix(rnorm(5000),100,50)
test[11:20,11:20] <- rnorm(100,3,0.3)

#Apply Plaid Biclustering
res <- biclust(test, method=BCPlaid())

#Apply default diagnosticTest
out <- diagnosticTest(BCresult=res, data=test, save_F=TRUE)

#Plot Distributions
diagnosticPlot2(out,binwidth=1)

}
\author{
Ewoud De Troyer
}
