\name{SyntrenEcoli}
\alias{SyntrenEcoli}
\alias{E. coli microarray synthetic data matrix}
\alias{Escherichia Coly data matrix generated by SynTReN}

\docType{data}

\title{SynTReN E. coli}
\description{Synthetic microarray data matrix generated by Syntren for 20 experiments
using 200 genes from Transcription Regulatory Network of Shen-Orr et al. (2002).}
\usage{
  data(SyntrenEcoli)
  }

\format{
  Data structure with information about the expression levels of 200 genes over 20 conditions.
  Conditions are named as C1... C20
}
%\details{}
%\value{}
\references{

@ARTICLE{Shen-Orr2002,
  author = {Shai S. Shen-Orr and Ron Milo and Shmoolik Mangan and Uri Alon},
  title = {Network motifs in the transcriptional regulation network of Escherichia
	coli},
  journal = {Nature Genetics},
  year = {2002},
  volume = {31},
  pages = {64--68}
  }
  
  @ARTICLE{Bulcke2006,
  author = {Tim Van den Bulcke and Koenraad Van Leemput and Bart Naudts and Piet
	van Remortel and Hongwu Ma and Alain Verschoren and Bart De Moor
	and Kathleen Marchal},
  title = {SynTReN: a generator of synthetic gene expression data for design
	and analysis of structure learning algorithms},
  journal = {BMC Bioinformatics},
  year = {2006},
  volume = {7},
  number = {43},
  doi = {10.1186/1471-2105-7-43},
  url = {http://www.biomedcentral.com/1471-2105/7/43}
  }
}


\source{SynTReN software can be downloaded at http://homes.esat.kuleuven.be/~kmarchal/SynTReN/index.html
}
%\seealso{
  %parallelCoordinates for alternative representation of biclusters.
%  \item{\code{\link{bubbleplot}}}{ for simultaneous representation of biclusters.}
%  }
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{datasets}
