\name{biclust-method}
\title{The biclust Method}
\alias{biclust}
\alias{biclust-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\description{Perfoms a choosen biclust algorithm }

\usage{
\S4method{biclust}{Biclust}  }

\author{Sebastian Kaiser 
\email{sebastian.kaiser@stat.uni-muenchen.de}
}

\seealso{
  \code{\link{biclust-method}}, \code{\link{BiclustMethod-class}}
}

\keyword{cluster}


