\name{biclust}
\title{The biclust Method}
\alias{biclust}
\alias{biclust-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\description{Perfoms a choosen biclust algorithm }

\usage{
\S4method{biclust}{Biclust}(x,method,...)  }

\arguments{
\item{x}{Data matrix.}
\item{method}{An object of class \code{"BiclustMethod"}.}
\item{\ldots}{Additional Parameters of the \code{"BiclustMethod"} }

\author{Sebastian Kaiser 
\email{sebastian.kaiser@stat.uni-muenchen.de}
}

\seealso{
  \code{\link{Biclust-class}}, \code{\link{BCCC}}, \code{\link{BCXmotifs}}, \code{\link{BCPlaid}}, \code{\link{BCSpectral}}, \code{\link{BCBimax}}, \code{\link{BiclustMethod-class}}
}

\keyword{cluster}


