\name{BiclustMethod-class}
\docType{class}
\title{The BiclustMethod Virtual Class}
\alias{BiclustMethod}
\alias{BiclustMethod-class}
%- Also NEED an '\alias' for EACH other topic documented here.
\description{BiclustMethod is the virtual class structure for algorithms provided in the package. In order to use the \code{biclust()} function a algorithm has to have a class inherit from here.}

\section{Algorithms}{
  Currently 6 classes inherit from BiclustMethod:
  \code{\link{BCCC}}, \code{\link{BCXmotifs}}, \code{\link{BCPlaid}}, \code{\link{BCSpectral}}, \code{\link{BCBimax}}, \code{\link{BCQuest}}
  }
  

\author{Sebastian Kaiser 
\email{sebastian.kaiser@stat.uni-muenchen.de}
}

\seealso{
  \code{\link{biclust}}, \code{\link{Biclust-class}}, \code{\link{BCCC}}, \code{\link{BCXmotifs}}, \code{\link{BCPlaid}}, \code{\link{BCSpectral}}, \code{\link{BCBimax}}, \code{\link{BCQuest}}, \code{\link{BiclustMethod-class}}
}
\keyword{classes}
%\examples{}
