\name{SyntrenEcoli}
\alias{SyntrenEcoli}
\alias{E. coli microarray synthetic data matrix}
\alias{Escherichia Coly data matrix generated by SynTReN}

\docType{data}

\title{SynTReN E. coli}
\description{Synthetic microarray data matrix generated by Syntren for 20 experiments
using 200 genes from Transcription Regulatory Network of Shen-Orr et al. (2002).}
\usage{
  data(SyntrenEcoli)
  }

\format{
  Data structure with information about the expression levels of 200 genes over 20 conditions.
  Conditions are named as C1... C20
}
%\details{}
%\value{}
\references{

  Shen-Orr et al., "Network motifs in the transcriptional regulation network of Escherichia coli",
  Nature Genetics 2002, volume 31, pages 64-68.

  
  Tim Van den Bulcke et al., "SynTReN: a generator of synthetic gene expression data for design
	and analysis of structure learning algorithms", BMC Bioinformatics, 2006,
  volume 7, number 43.

}


\source{SynTReN software can be downloaded at http://homes.esat.kuleuven.be/~kmarchal/SynTReN/index.html
}
%\seealso{
  %parallelCoordinates for alternative representation of biclusters.
%  \item{\code{\link{bubbleplot}}}{ for simultaneous representation of biclusters.}
%  }
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{datasets}
